refactor: BunchOfLottoTest 추가
